requirements.txt
Django==2.1.11
django-crispy-forms==1.7.2
Pillow==6.2.0
pytz==2018.4
djangorestframework==3.11.0
pytest==5.3.5
mixer==6.1.3
pytest-cov==2.8.1